% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes_table}
\alias{barcodes_table}
\title{Barcode Reference Sequences}
\format{A data frame with 500 rows and 2 variables:
\describe{
  \item{Labels}{Barcode Sequence IDs}
  \item{Code}{DNA Sequences of length 18}
}}
\source{
Randomly generated sequences
}
\usage{
barcodes_table
}
\description{
A dataset containing identifiers of 500 barcode sequences.
}
\keyword{datasets}
