% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probability.R
\name{calculate_probability}
\alias{calculate_probability}
\title{Calculate Joint Probability of Map}
\usage{
calculate_probability(reference, read, phred_score)
}
\arguments{
\item{reference}{The barcode sequence to which \code{read} maps to. Should be the same length as \code{read}.}

\item{read}{The read sequence which maps to \code{barcode}. Should be the same length as \code{barcode}.}

\item{phred_score}{A string of Phred scores. Each character is a representation of the probability that
the corresponding base call is an error.}
}
\value{
Returns a scalar joint probability, which is the product of marginal probabilities obtained from
\code{phred_to_prob}
}
\description{
\code{calculate_probability} takes a read and reference of equal length \code{n}, finds the positions of mismatch,
then calculates the joint probability of the read mapping to the reference based on the given phred score.
For a given position in a mapping and \code{P} being the probability of error at that position, the probability
of mapping is \eqn{1 - P} if the calls are the same, and \eqn{P/3} if the calls are different.
If there are no mismatches, the joint probability is \eqn{(1 - P)^n}.
}
