% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_barcode_table.R
\name{get_barcode_table}
\alias{get_barcode_table}
\title{Get Barcode Dataframe}
\usage{
get_barcode_table()
}
\value{
Returns a dataframe of barcode identifiers and reference barcodes
}
\description{
\code{get_barcode_table} reads in the fasta file that is included with this package.
}
\examples{
barcodes_df <- get_barcode_table()
}
