% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_of_optimal_mapping.R
\name{get_index_of_optimal_mapping}
\alias{get_index_of_optimal_mapping}
\title{Calculate Joint Probability of Map}
\usage{
get_index_of_optimal_mapping(barcodes, read, read_phred, filtered_barcodes)
}
\arguments{
\item{read}{The read sequence which maps to \code{barcode}. Should be the same length as \code{barcode}.}

\item{reference}{The barcode sequence to which \code{read} maps to. Should be the same length as \code{read}.}
}
\value{
Returns a scalar joint probability
}
\description{
\code{calculate_probability} takes a read and reference of equal length \code{n}, finds the positions of mismatch,
then calculates the joint probability of the read mapping to the reference based on the given phred score.
For a given position in a mapping and \code{P} being the probability of error at that position, the probability
of mapping is \eqn{1 - P} if the calls are the same, and \eqn{P/3} if the calls are different.
If there are no mismatches, the joint probability is \eqn{(1 - P)^n}.
}
\examples{
phred_to_prob('BEEEEFEFEEEFBEEDEE')
phred_to_prob('989BBAB7A3E667BAA?')
}
