% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapped_barcode_indices.R
\name{get_mapped_barcode_indices}
\alias{get_mapped_barcode_indices}
\title{Find Indices of Potential Barcode Mappings}
\usage{
get_mapped_barcode_indices(barcodes, read, distance_threshold = 2)
}
\arguments{
\item{barcodes}{Character vector of reference sequences}

\item{read}{Character string with the same length as each barcode in \code{barcodes}}

\item{distance_threshold}{Integer specifying the Hamming distance threshold for which we want to filter barcodes}
}
\value{
Returns a vector of integers which are the indices of barcodes within 2 Hamming distance of the read
}
\description{
\code{get_mapped_barcode_indices} takes a vector of barcodes and a single read, then
returns the indices of the barcodes that were within two hamming distance units
from the read.
}
