% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reads_table.R
\name{get_reads_table}
\alias{get_reads_table}
\title{Get Reads Dataframe}
\usage{
get_reads_table()
}
\value{
Returns a dataframe of sample identifiers, the called reads, and their phred scores
}
\description{
\code{get_reads_table} reads in the fastq file that is included with this package.
}
\examples{
reads_df <- get_reads_table()
}
