% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming}
\alias{hamming}
\title{Compute Hamming Distance}
\usage{
hamming(str1, str2)
}
\arguments{
\item{str1}{The first string, should be equal length to \code{str2}}

\item{str2}{The second string, should be equal length to \code{str1}}
}
\value{
Returns the hamming distance, a scalar integer
}
\description{
\code{hamming} computes the hamming distance between two equal length strings. The hamming distance
is the number of different characters between strings. Is called by \code{\link{get_mapped_barcode_indices}}.
}
