% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reads.R
\name{map_reads}
\alias{map_reads}
\title{Map Reads to Reference Barcodes}
\usage{
map_reads(barcodes_table, reads_table)
}
\arguments{
\item{barcodes_table}{The full table of reference barcodes, which contains a column titled \code{Labels},
which contains the sequence ID, and another called \code{Code}, which contains the barcode sequence.}

\item{reads_table}{The full table of reads, which contains a column titled \code{Read_ID}, another called
\code{Called_Read}, and a third column titled \code{Phred_Score}.}
}
\value{
Returns the original reads table along with the barcode that was mapped and the probability of
the map.
}
\description{
\code{map_reads} is the primary workhorse of this package. It iterates through the reads_table,
finds the barcodes that are within 2 Hamming distances, and then computes the probability of
the read mapping to each barcode that is within 2 Hamming distances. Then, out of this list
of probabilities, it selects the first barcode with the greatest probability of being a match.
}
\examples{
mapped_reads <- map_reads(reads_table = reads_table, barcodes_table = barcodes_table)
mapped_reads \%>\% group_by(mapped_barcodes) \%>\%
     summarize(Number_Mapped = n(), Proportion_Mapped = n()/nrow(mapped_reads))
}
