% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phred_to_prob.R
\name{phred_to_prob}
\alias{phred_to_prob}
\title{Convert Phred Score to Probability}
\usage{
phred_to_prob(phred_score)
}
\arguments{
\item{phred_score}{A string of Phred scores. Each character is a representation of the probability that
the corresponding base call is an error.}
}
\value{
Returns a numeric vector of probabilities with length equal to the number of characters given in
\code{phred_score}
}
\description{
\code{phred_to_prob} converts a series of \href{https://en.wikipedia.org/wiki/Phred_quality_score}{phred scores},
which are ASCII characters used to encode the probabilities of a sequencing error for given base calls
in a sequence, to a vector of probability values.
\code{phred_to_prob} is called by \code{\link{calculate_probability}}.
}
